auth util package path refactoring
