+++ src/main/java/kr/pincoin/durian/auth/util/jwt/JwtAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package kr.pincoin.durian.auth.jwt;
+package kr.pincoin.durian.auth.util.jwt;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
